smartroute-worker/requirements.txt
azure-functions
azure-storage-queue
psycopg2-binary